Finalizando modulo de cursos

+++ src/pages/Course/List/index.tsx
@@ -11,26 +11,42 @@ import CreateIcon from '@material-ui/icons/Create';
 
 const ListCourse = (): JSX.Element => {
   const handleDelete = async (id?: string) => {
-    try{
-      const currentSession = await Auth.currentSession();
-      await api.delete(`/course/${id}`, {
-        headers: {
-          'X-Cognito-ID-Token': currentSession.getIdToken().getJwtToken()
-        }
-      });
+    const currentSession = await Auth.currentSession();
 
-      Swal.fire({
-        icon: 'success',
-        title: 'Sucesso!',
-        text: 'O curso foi excluído com sucesso!'
-      });
-    } catch (error) {
-      Swal.fire({
-        icon: 'error',
-        title: error.response.data.title,
-        text: error.response.data.text,
-      });
-    }
+    Swal.fire({
+      icon: 'question',
+      title: 'Atenção!',
+      text: 'Deseja realmente excluir o curso selecionado?',
+      showCancelButton: true,
+      confirmButtonText: "Sim, confirmar!",
+      cancelButtonText: "Não!"
+    }).then( async (result) => {
+      if(result.isConfirmed){
+        try{
+          await api.delete(`/course/${id}`, {
+            headers: {
+              'X-Cognito-ID-Token': currentSession.getIdToken().getJwtToken()
+            }
+          });
+    
+          Swal.fire({
+            icon: 'success',
+            title: 'Sucesso!',
+            text: 'O curso foi excluído com sucesso!'
+          }).then(result => {
+            window.location.reload(false);
+          });
+        } catch(error) {
+          Swal.fire({
+            icon: 'error',
+            title: error.response.data.title,
+            text: error.response.data.text,
+          });
+        }
+      } else {
+        Swal.close();
+      }
+    });
   }
 
   return (
@@ -62,14 +78,16 @@ const ListCourse = (): JSX.Element => {
               link: "/course",
               title: "Visualizar",
               handle: () => {},
-              icon: <VisibilityIcon />
+              icon: <VisibilityIcon />,
+              action: "view"
             },
             {
               type: "link",
               link: "/course",
               title: "Editar",
               handle: () => {},
-              icon: <CreateIcon />
+              icon: <CreateIcon />,
+              action: "edit"
             },
             {
               type: "button",
